Strip known engine extensions from source list for manifests
Mincer resolves the source for an asset by itself when compiling to
a manifest file. So instead of source.css.less it expects source.css.
This conflicts with grunt which by default removes all files which can
not be found while globbing, so the final destination source.css can't
be supplied without specifying `nonull: true` in the options.

This works around the issue by resolving the filename through stripping
all file extensions belonging to a mincer supported engine and passing
the resulting file array to Manifest.compile()